add generating response loader and scroll gradient
