+++ assets/package.json
@@ -101,6 +101,7 @@
     "react-file-icon": "1.3.0",
     "react-file-picker": "0.0.6",
     "react-highlight.js": "1.0.7",
+    "react-merge-refs": "2.1.1",
     "react-portal": "4.2.2",
     "react-pulse-dot": "2.1.0",
     "react-resize-detector": "9.1.0",
